Add ability to provide header comments, added parsing of 'dns-search' (#24)
* Add ability to provide header comments

Added param to InterfacesWriter to add the ability to
publish a commented text string at the beginning of the
interfaces file.
Updated InterfaceReader to read in the first block of comments
at the beginning of the file as a header comment.

* Add configuration param 'dns-search'

Adding full read/write pipeline for interfaces 'dns-search'
option. Allows setting of custom dns search paths for private
dns servers.